Prøver å konvertere lifespan til int

+++ config/pureservice.php
@@ -15,7 +15,7 @@ return [
     ],
     'computer' => [
         'displayName' => env('PURESERVICE_COMPUTER_ASSETTYPE_NAME', 'Datamaskin'),
-        'lifespan' => env('PURESERVICE_COMPUTER_LIFESPAN', 4), // Forventet levetid, oppgitt som år
+        'lifespan' => (int)env('PURESERVICE_COMPUTER_LIFESPAN', 4), // Forventet levetid, oppgitt som år
         'asset_type_id' => null, // Settes automatisk basert på displayName
         'relationship_type_id' => null, // Settes automatisk basert på displayName
         'className' => '', // Settes automatisk basert på displayName
@@ -46,7 +46,7 @@ return [
     ],
     'mobile' => [
         'displayName' => env('PURESERVICE_MOBILE_ASSETTYPE_NAME', 'Mobilenhet'),
-        'lifespan' => env('PURESERVICE_MOBILE_LIFESPAN', 3), // Forventet levetid, oppgitt som år
+        'lifespan' => (int)env('PURESERVICE_MOBILE_LIFESPAN', 3), // Forventet levetid, oppgitt som år
         'asset_type_id' => null, // Settes automatisk basert på displayName
         'relationship_type_id' => null, // Settes automatisk basert på displayName
         'className' => '', // Settes automatisk basert på displayName
